+++ api/src/index.mjs
@@ -3,6 +3,7 @@ import express from 'express'
 import { wrapExpress } from '#shared/utils'
 import { getPreset } from '#config'
 import cookieParser from 'cookie-parser'
+import bodyParser from 'body-parser'
 // Routes
 import { routes } from '#routes/index'
 // Bootstrap configuration
@@ -32,6 +33,11 @@ app.use(guardRoutes)
  * Add support for cookies with a limit to the request body
  */
 app.use(cookieParser())
+
+/*
+ * Add support for form bodies (required for OIDC flow)
+ */
+app.use(bodyParser.urlencoded())
 
 /*
  * Add custom middleware to load roles from header
@@ -114,7 +120,7 @@ export async function reload() {
   /*
    * This does the actual reloading
    */
-  await reloadConfiguration()
+  await reloadConfiguration(app)
 
   /*
    * Let the world know we are ready