+++ ui/components/layout/oidc.mjs
@@ -1,0 +1,3 @@
+export const OidcLayout = ({ children }) => (
+  <div className="w-full m-auto p-0" style={{ minHeight: "calc(100vh - 8rem)"}}>{children}</div>
+)